Revert "Merge pull request #216 from judaschwartz/patch-5"
This reverts commit d3608f399fb5df0b90f5dc427aabada5716bbfd4, reversing
changes made to ebd91a0ac192d487fad8b8c6efcc432d0e5a08f3.

As it turns out, many of these changes were dubious, such as
"permission" or "environment".

Also, we need to decide if this is a breaking change or not.
In the meantime, let us revert this.